requirements.txt
anthropic>=0.8.0
supabase-py>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
python-dateutil>=2.8.2
requests>=2.31.0
